Rename app to Control de Loterías and fix hour HHMM format
- Window title, sidebar and assembly name changed to "Control de Loterías"
- CambiarHora now takes int; AplicarHora sends horas*100+minutos (HHMM format)
- SceneController rewritten to match reference project (toggle + direct methods)
- Hour slider fixed: display uses /60, presets use *60

+++ LoteriaTwo/Services/SceneController.cs
@@ -99,7 +99,7 @@ public class SceneController
 
     // ── Propiedades ─────────────────────────────────────────────────────────────
 
-    public async void CambiarHora(float hora)
+    public async void CambiarHora(int hora)
         => await _signals.EnviarPropiedadAsync(_currentLevel!, "Hora", hora);
 
     public async void CambiarHoraAutomatica(bool activar)
